+++ Content.Server/Bed/Cryostorage/CryostorageSystem.cs
@@ -27,10 +27,6 @@ using Robust.Shared.Containers;
 using Robust.Shared.Enums;
 using Robust.Shared.Network;
 using Robust.Shared.Player;
-using Content.Shared.Audio;
-using Robust.Shared.Audio.Systems;
-using Content.Shared.DoAfter;
-using Content.Shared.SS220.CryopodSSD;
 using Content.Server.Forensics;
 using Robust.Shared.Utility;
 
@@ -55,8 +51,6 @@ public sealed class CryostorageSystem : SharedCryostorageSystem
     [Dependency] private readonly StationRecordsSystem _stationRecords = default!;
     [Dependency] private readonly TransformSystem _transform = default!;
     [Dependency] private readonly UserInterfaceSystem _ui = default!;
-    [Dependency] private readonly SharedAudioSystem _audioSystem = default!;
-    [Dependency] private readonly SharedDoAfterSystem _doAfterSystem = default!;
 
     /// <inheritdoc/>
     public override void Initialize()
@@ -65,7 +59,6 @@ public sealed class CryostorageSystem : SharedCryostorageSystem
 
         SubscribeLocalEvent<CryostorageComponent, BeforeActivatableUIOpenEvent>(OnBeforeUIOpened);
         SubscribeLocalEvent<CryostorageComponent, CryostorageRemoveItemBuiMessage>(OnRemoveItemBuiMessage);
-        SubscribeLocalEvent<CryostorageComponent, TeleportToCryoFinished>(OnTeleportFinished);
 
         SubscribeLocalEvent<CryostorageContainedComponent, PlayerSpawnCompleteEvent>(OnPlayerSpawned);
         SubscribeLocalEvent<CryostorageContainedComponent, MindRemovedMessage>(OnMindRemoved);
@@ -244,14 +237,14 @@ public sealed class CryostorageSystem : SharedCryostorageSystem
 
             // _stationRecords.RemoveRecord(key, stationRecords);
 
-            // start 220 cryo department
+            // start 220 cryo department record
             var recordPairTry = FindEntityStationRecordKey(station.Value, ent);
             if (recordPairTry is { } recordPair)
             {
                 recordPair.Item2.IsInCryo = true;
                 _stationRecords.Synchronize(station.Value);
             }
-            // end 220 cryo department
+            // end 220 cryo department record
         }
 
         _chatSystem.DispatchStationAnnouncement(station.Value,
@@ -366,77 +359,7 @@ public sealed class CryostorageSystem : SharedCryostorageSystem
         }
     }
 
-    // START 220 TELEPORT TO CRYO
-    /// <summary>
-    /// Tries to teleport target inside cryopod, if any available
-    /// </summary>
-    /// <param name="target"> Target to teleport in first matching cryopod</param>
-    /// <returns> true if player successfully transferred to cryo storage, otherwise returns false</returns>
-    public bool TeleportEntityToCryoStorageWithDelay(EntityUid target)
-    {
-        var station = _station.GetOwningStation(target);
-
-        if (station is null)
-            return false;
-
-        foreach (var comp in EntityQuery<CryostorageComponent>())
-        {
-            if (comp.StoredPlayers.Contains(target))
-                return true;
-        }
-
-        var cryopodSSDComponents = EntityQueryEnumerator<CryostorageComponent>();
-
-        while (cryopodSSDComponents.MoveNext(out var cryopodSSDUid, out var cryopodSSDComp))
-        {
-            // if (cryopodSSDComp.BodyContainer.ContainedEntity is null
-            if (!cryopodSSDComp.StoredPlayers.Contains(target)  // todo check
-            && _station.GetOwningStation(cryopodSSDUid) == station)
-            {
-                var portal = Spawn("CryoStoragePortal", Transform(target).Coordinates);
-
-                if (TryComp<AmbientSoundComponent>(portal, out var ambientSoundComponent))
-                {
-                    _audioSystem.PlayPvs(ambientSoundComponent.Sound, portal);
-                }
-
-                var doAfterArgs = new DoAfterArgs(EntityManager, target, TimeSpan.FromSeconds(4f), new TeleportToCryoFinished(GetNetEntity(portal)), cryopodSSDUid) // todo edit TimeSpan.FromSeconds(4)
-                {
-                    BreakOnDamage = false,
-                    BreakOnMove = false,
-                    NeedHand = false,
-                };
-
-                if (!_doAfterSystem.TryStartDoAfter(doAfterArgs))
-                    QueueDel(portal);
-
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private void OnTeleportFinished(Entity<CryostorageComponent> ent, ref TeleportToCryoFinished args)
-    {
-        if (_container.TryGetContainer(ent.Owner, ent.Comp.ContainerId, out var container))
-        {
-            AdminLog.Add(LogType.CryoStorage, LogImpact.High,
-                $"{ToPrettyString(args.User):player} was teleported to cryostorage {ToPrettyString(ent)}");
-            _container.Insert(args.User, container);
-        }
-
-        if (TryComp<CryostorageContainedComponent>(args.User, out var contained))
-            contained.GracePeriodEndTime = TimeSpan.Zero;
-
-        var portalEntity = GetEntity(args.PortalId);
-
-        if (TryComp<AmbientSoundComponent>(portalEntity, out var ambientSoundComponent))
-            _audioSystem.PlayPvs(ambientSoundComponent.Sound, portalEntity);
-
-        EntityManager.DeleteEntity(portalEntity);
-    }
-
+    // start 220 cryo department record
     private (StationRecordKey, GeneralStationRecord)? FindEntityStationRecordKey(EntityUid station, EntityUid uid)
     {
         if (TryComp<DnaComponent>(uid, out var dnaComponent))
@@ -451,5 +374,5 @@ public sealed class CryostorageSystem : SharedCryostorageSystem
 
         return null;
     }
-    // END 220 TELEPORT TO CRYO
+    // end 220 cryo department record
 }
