+++ Content.Shared/SS220/TeleportAFKtoCryoSystem/SharedTeleportAFKtoCryoSystem.cs
@@ -1,7 +1,7 @@
 using Robust.Shared.Serialization;
 using Content.Shared.DoAfter;
 
-namespace Content.Shared.SS220.CryopodSSD;
+namespace Content.Shared.SS220.TeleportAFKtoCryoSystem;
 
 [Serializable, NetSerializable]
 public sealed partial class TeleportToCryoFinished : SimpleDoAfterEvent
